requirements.txt
aiocryptopay==0.4.8
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosend==2.1.2
aiosignal==1.4.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==2.1.1
colorama==0.4.6
configparser==6.0.0
console==0.9907
cryptography==45.0.5
ezenv==0.92
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinxed==1.3.0
lxml==6.0.0
magic-filter==1.0.12
multidict==6.6.4
mutf8==1.0.6
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pyCraft @ git+https://github.com/ammaraskar/pyCraft.git@bcd156e83bd35fb2fdc6cfa64f97bfc96d7cc1a1
pydantic==2.11.7
pydantic_core==2.33.2
PyNBT==3.1.0
PySocks==1.7.1
pyTelegramBotAPI==4.15.2
python-dotenv==1.1.1
python-telegram-bot==22.3
readchar==4.0.5
requests==2.28.1
requests-toolbelt==0.10.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
stem==1.8.2
StrEnum==0.4.15
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
yarl==1.20.1
